+++ housing1000/Housing1000_iOS/Housing1000/NSData+MD5.m
@@ -14,7 +14,7 @@
 - (NSString *)generateMD5Hash
 {
     unsigned char md5Buffer[CC_MD5_DIGEST_LENGTH];
-    CC_MD5(self.bytes, self.length, md5Buffer);
+    CC_MD5(self.bytes, (int)self.length, md5Buffer);
     NSMutableString *output = [NSMutableString stringWithCapacity:CC_MD5_DIGEST_LENGTH * 2];
     for(int i = 0; i < CC_MD5_DIGEST_LENGTH; i++)
         [output appendFormat:@"%02x",md5Buffer[i]];
